taint/test/mochitest/test_navigator.html

<!DOCTYPE HTML>
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <meta charset="utf-8">
    <title>Check Navigator Sink</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script "text/javascript">


      let sink_names = [
          "navigator.sendBeacon(url)",
          "navigator.sendBeacon(body)",
      ];

      function check_tainted(str) {
          SimpleTest.ok(str.taint.length > 0, "Check tainted");
      }

      function check_untainted(str) {
          SimpleTest.ok(str.taint.length == 0, "Check untainted");
      }

      let i = 0;
      SimpleTest.waitForExplicitFinish();
      addEventListener("__taintreport", (report) => {
          let flow = report.detail.str.taint[0].flow;
          SimpleTest.is(flow[1].operation, sink_names[i]);

          i += 1;
          if (i >= sink_names.length) {
              SimpleTest.finish();
          }
      }, false);

      function startTest() {
          let navigator = window.navigator;
          navigator.sendBeacon(String.tainted("url"));
          navigator.sendBeacon("url", String.tainted("body"));
      }

    </script>
  </head>

  <body onload="startTest();">

</html>
